refactor(user): split userController into profile / deletion / tags-badges / blocking controllers
Break the 2,188-line userController.js into focused controllers by
responsibility, following the teamController pattern (separately named
controllers that userRoutes imports directly). Verbatim moves, no behavior
change; the public route interface is unchanged.

- userController.js (2,188 -> 638): profile CRUD only (getUsers, getUserById,
  updateUser, deleteAvatar).
- userDeletionController.js (new): deleteUser + deletionPreview plus their
  deletion helpers (logDeletionPhase, insertNotificationRecord,
  getSuccessorCandidatesByTeam, DELETED_USER_DISPLAY_NAME).
- userTagsBadgesController.js (new): getUserTags, updateUserTags,
  updateUserBadgeVisibility, getUserBadges.
- userBlockingController.js (new): getBlockedUsers, blockUser, unblockUser,
  getBlockRelationships plus isSelf/emitBlocksUpdated.
- src/utils/user/userControllerHelpers.js (new): shared pure helpers
  buildUserDisplayName and toIsoString.

userRoutes.js now imports the three new controllers directly. Also removed two
imports that became unused in userController (bcrypt, the shared helpers) and
the orphaned block-section comments. The two account-deletion test files now
require userDeletionController. README controllers tree updated.
