+++ src/controllers/userController.js
@@ -1,6 +1,5 @@
 const db = require("../config/database");
 const { pool } = db;
-const bcrypt = require("bcrypt");
 const {
   hasLocationChanged,
   resolveLocationData,
@@ -21,121 +20,12 @@ const SAFE_PUBLIC_USER_KEYS = [
   'is_synthetic', 'created_at', 'tags',
 ];
 
-const DELETED_USER_DISPLAY_NAME = "Former Lomir User";
-
 const sanitizePublicUser = (row) => {
   const safe = {};
   for (const key of SAFE_PUBLIC_USER_KEYS) {
     if (row[key] !== undefined) safe[key] = row[key];
   }
   return safe;
-};
-
-const buildUserDisplayName = (user) => {
-  const fullName = [user.first_name, user.last_name].filter(Boolean).join(" ");
-  return fullName || user.username;
-};
-
-const toIsoString = (value) => {
-  if (value instanceof Date) {
-    return value.toISOString();
-  }
-
-  return value;
-};
-
-const logDeletionPhase = (phase, details) => {
-  if (process.env.NODE_ENV === "production") {
-    return;
-  }
-
-  if (details !== undefined) {
-    console.log(`[deleteUser] ${phase}`, details);
-    return;
-  }
-
-  console.log(`[deleteUser] ${phase}`);
-};
-
-const insertNotificationRecord = async (
-  client,
-  {
-    userId,
-    type,
-    title,
-    message = null,
-    referenceType = null,
-    referenceId = null,
-    teamId = null,
-    actorId = null,
-  },
-) => {
-  await client.query(
-    `INSERT INTO notifications (user_id, type, title, message, reference_type, reference_id, team_id, actor_id)
-     VALUES ($1, $2, $3, $4, $5, $6, $7, $8)`,
-    [
-      userId,
-      type,
-      title,
-      message,
-      referenceType,
-      referenceId,
-      teamId,
-      actorId,
-    ],
-  );
-};
-
-const getSuccessorCandidatesByTeam = async (queryable, teamIds, excludedUserId) => {
-  if (teamIds.length === 0) {
-    return new Map();
-  }
-
-  const result = await queryable.query(
-    `
-    SELECT
-      tm.team_id,
-      tm.user_id,
-      tm.role,
-      tm.joined_at,
-      u.first_name,
-      u.last_name,
-      u.username
-    FROM team_members tm
-    JOIN users u ON u.id = tm.user_id
-    WHERE tm.team_id = ANY($1::int[])
-      AND tm.user_id != $2
-      AND tm.role IN ('admin', 'member')
-    ORDER BY
-      tm.team_id ASC,
-      CASE
-        WHEN tm.role = 'admin' THEN 0
-        WHEN tm.role = 'member' THEN 1
-        ELSE 2
-      END,
-      tm.joined_at ASC NULLS LAST,
-      tm.user_id ASC
-    `,
-    [teamIds, excludedUserId],
-  );
-
-  const candidatesByTeamId = new Map();
-
-  for (const row of result.rows) {
-    const teamId = Number(row.team_id);
-    const existingCandidates = candidatesByTeamId.get(teamId) || [];
-
-    existingCandidates.push({
-      userId: Number(row.user_id),
-      name: buildUserDisplayName(row),
-      role: row.role,
-      joinedAt: toIsoString(row.joined_at),
-    });
-
-    candidatesByTeamId.set(teamId, existingCandidates);
-  }
-
-  return candidatesByTeamId;
 };
 
 /**
@@ -740,1449 +630,9 @@ const deleteAvatar = async (req, res) => {
   }
 };
 
-/**
- * @description Delete a user's account
- * @route DELETE /api/users/:id
- * @access Private (Requires authentication and authorization)
- */
-const deleteUser = async (req, res) => {
-  let client = null;
-  let transactionOpen = false;
-  let avatarUrl = null;
-  let avatarFileId = null;
-  let teamIdsForSockets = [];
-  let dmPartnerIds = [];
-  let ownershipTransferEvents = [];
-  let reopenedRoleEvents = [];
-
-  try {
-    const userId = parseInt(req.params.id, 10);
-    const { password, ownershipOverrides = [] } = req.body || {};
-
-    // Verify the user making the request is the same as the user being deleted
-    if (Number(req.user.id) !== userId) {
-      return res.status(403).json({
-        success: false,
-        message: "You can only delete your own account",
-      });
-    }
-
-    if (!password) {
-      return res.status(400).json({
-        success: false,
-        message: "Password is required",
-      });
-    }
-
-    if (!Array.isArray(ownershipOverrides)) {
-      return res.status(400).json({
-        success: false,
-        message: "ownershipOverrides must be an array",
-      });
-    }
-
-    const ownershipOverrideMap = new Map();
-
-    for (const override of ownershipOverrides) {
-      const teamId = Number(override?.teamId);
-      const successorId = Number(override?.successorId);
-
-      if (!Number.isInteger(teamId) || !Number.isInteger(successorId)) {
-        return res.status(400).json({
-          success: false,
-          message: "Each ownership override must include teamId and successorId",
-        });
-      }
-
-      ownershipOverrideMap.set(teamId, successorId);
-    }
-
-    client = await db.pool.connect();
-
-    const rollbackAndRespond = async (status, payload) => {
-      if (transactionOpen) {
-        await client.query("ROLLBACK");
-        transactionOpen = false;
-      }
-
-      return res.status(status).json(payload);
-    };
-
-    await client.query("BEGIN");
-    transactionOpen = true;
-
-    logDeletionPhase("Phase A - gather context", { userId });
-
-    const userResult = await client.query(
-      `SELECT id, first_name, last_name, username, avatar_url, avatar_file_id, password_hash
-       FROM users
-       WHERE id = $1`,
-      [userId],
-    );
-
-    if (userResult.rows.length === 0) {
-      return rollbackAndRespond(404, {
-        success: false,
-        message: "User not found",
-      });
-    }
-
-    const user = userResult.rows[0];
-    const passwordMatches = await bcrypt.compare(password, user.password_hash);
-
-    if (!passwordMatches) {
-      return rollbackAndRespond(401, {
-        success: false,
-        message: "Password is incorrect",
-      });
-    }
-
-    avatarUrl = user.avatar_url;
-    avatarFileId = user.avatar_file_id;
-
-    const fullName = [user.first_name, user.last_name].filter(Boolean).join(" ");
-    const userDisplayName = fullName || user.username;
-
-    const [
-      membershipsResult,
-      ownedTeamsResult,
-      filledRolesResult,
-      dmPartnersResult,
-    ] = await Promise.all([
-      client.query(
-        `
-        SELECT
-          tm.team_id,
-          tm.role,
-          tm.joined_at,
-          t.name AS team_name
-        FROM team_members tm
-        JOIN teams t ON t.id = tm.team_id
-        WHERE tm.user_id = $1
-        ORDER BY t.name ASC, tm.team_id ASC
-        `,
-        [userId],
-      ),
-      client.query(
-        `
-        SELECT
-          t.id AS team_id,
-          t.name AS team_name,
-          COUNT(tm_all.user_id)::int AS member_count,
-          (COUNT(tm_all.user_id) FILTER (WHERE tm_all.user_id != $1))::int AS other_member_count
-        FROM teams t
-        JOIN team_members tm_owner
-          ON tm_owner.team_id = t.id
-         AND tm_owner.user_id = $1
-         AND tm_owner.role = 'owner'
-        JOIN team_members tm_all ON tm_all.team_id = t.id
-        GROUP BY t.id, t.name
-        ORDER BY t.name ASC, t.id ASC
-        `,
-        [userId],
-      ),
-      client.query(
-        `
-        SELECT
-          vr.id AS role_id,
-          vr.role_name,
-          vr.team_id,
-          t.name AS team_name
-        FROM team_vacant_roles vr
-        JOIN teams t ON t.id = vr.team_id
-        WHERE vr.filled_by = $1
-          AND vr.status = 'filled'
-        ORDER BY t.name ASC, vr.role_name ASC, vr.id ASC
-        `,
-        [userId],
-      ),
-      client.query(
-        `
-        SELECT DISTINCT
-          CASE
-            WHEN sender_id = $1 THEN receiver_id
-            ELSE sender_id
-          END AS partner_id
-        FROM messages
-        WHERE team_id IS NULL
-          AND (sender_id = $1 OR receiver_id = $1)
-          AND (
-            CASE
-              WHEN sender_id = $1 THEN receiver_id
-              ELSE sender_id
-            END
-          ) IS NOT NULL
-        `,
-        [userId],
-      ),
-    ]);
-
-    const memberships = membershipsResult.rows.map((row) => ({
-      teamId: Number(row.team_id),
-      teamName: row.team_name,
-      role: row.role,
-      joinedAt: toIsoString(row.joined_at),
-    }));
-
-    const ownedTeams = ownedTeamsResult.rows.map((row) => ({
-      teamId: Number(row.team_id),
-      teamName: row.team_name,
-      memberCount: Number(row.member_count),
-      otherMemberCount: Number(row.other_member_count),
-    }));
-
-    const teamsToDelete = ownedTeams.filter((team) => team.otherMemberCount === 0);
-    const teamsToTransfer = ownedTeams.filter((team) => team.otherMemberCount > 0);
-    const teamsToTransferIdSet = new Set(
-      teamsToTransfer.map((team) => team.teamId),
-    );
-
-    const invalidOverrideTeamIds = Array.from(ownershipOverrideMap.keys()).filter(
-      (teamId) => !teamsToTransferIdSet.has(teamId),
-    );
-
-    if (invalidOverrideTeamIds.length > 0) {
-      return rollbackAndRespond(400, {
-        success: false,
-        message:
-          "Ownership overrides can only be provided for teams that require ownership transfer",
-      });
-    }
-
-    const filledRoles = filledRolesResult.rows.map((row) => ({
-      roleId: Number(row.role_id),
-      roleName: row.role_name,
-      teamId: Number(row.team_id),
-      teamName: row.team_name,
-    }));
-
-    teamIdsForSockets = Array.from(
-      new Set(memberships.map((membership) => membership.teamId)),
-    );
-    dmPartnerIds = dmPartnersResult.rows
-      .map((row) => Number(row.partner_id))
-      .filter((partnerId) => Number.isInteger(partnerId));
-
-    logDeletionPhase("Phase B - messages and chat cleanup", {
-      teamCount: memberships.length,
-      dmPartnerCount: dmPartnerIds.length,
-    });
-
-    await client.query(
-      `DELETE FROM messages
-       WHERE (sender_id = $1 OR receiver_id = $1)
-         AND team_id IS NULL`,
-      [userId],
-    );
-
-    for (const membership of memberships) {
-      await client.query(
-        `INSERT INTO messages (sender_id, team_id, content, sent_at)
-         VALUES ($1, $2, $3, NOW())`,
-        [
-          null,
-          membership.teamId,
-          `🚪 ${DELETED_USER_DISPLAY_NAME} has left Lomir.`,
-        ],
-      );
-    }
-
-    await client.query(
-      `
-      UPDATE messages
-      SET content = CASE
-        WHEN $2 <> ''
-          THEN REPLACE(REPLACE(content, $2, 'Former Lomir User'), $3, 'Former Lomir User')
-        ELSE REPLACE(content, $3, 'Former Lomir User')
-      END
-      WHERE sender_id = $1
-        AND team_id IS NOT NULL
-        AND (
-          content LIKE '%👋%'
-          OR content LIKE '%🚪%'
-          OR content LIKE '%👑%'
-          OR content LIKE '%🎯%'
-          OR content LIKE '%✅%'
-          OR content LIKE '%❌%'
-          OR content LIKE '%🎉%'
-          OR content LIKE '%🔓%'
-        )
-      `,
-      [userId, fullName, user.username],
-    );
-
-    logDeletionPhase("Phase C - team ownership cleanup", {
-      teamsToDelete: teamsToDelete.length,
-      teamsToTransfer: teamsToTransfer.length,
-    });
-
-    const deletedTeamIds = new Set();
-
-    for (const team of teamsToDelete) {
-      const dissolutionTitle = `The team ${team.teamName} has been dissolved`;
-
-      await client.query(
-        `
-        UPDATE badge_awards
-        SET custom_team_name = (SELECT name FROM teams WHERE id = $1),
-            team_id = NULL
-        WHERE team_id = $1
-          AND team_id IS NOT NULL
-        `,
-        [team.teamId],
-      );
-
-      const pendingApplicantsResult = await client.query(
-        `
-        SELECT DISTINCT applicant_id
-        FROM team_applications
-        WHERE team_id = $1
-          AND status = 'pending'
-        `,
-        [team.teamId],
-      );
-
-      for (const applicant of pendingApplicantsResult.rows) {
-        await insertNotificationRecord(client, {
-          userId: Number(applicant.applicant_id),
-          type: "team_dissolved",
-          title: dissolutionTitle,
-          actorId: userId,
-        });
-      }
-
-      const pendingInviteesResult = await client.query(
-        `
-        SELECT DISTINCT invitee_id
-        FROM team_invitations
-        WHERE team_id = $1
-          AND status = 'pending'
-        `,
-        [team.teamId],
-      );
-
-      for (const invitee of pendingInviteesResult.rows) {
-        await insertNotificationRecord(client, {
-          userId: Number(invitee.invitee_id),
-          type: "team_dissolved",
-          title: dissolutionTitle,
-          actorId: userId,
-        });
-      }
-
-      await client.query("DELETE FROM team_invitations WHERE team_id = $1", [
-        team.teamId,
-      ]);
-      await client.query("DELETE FROM team_applications WHERE team_id = $1", [
-        team.teamId,
-      ]);
-      await client.query("DELETE FROM messages WHERE team_id = $1", [team.teamId]);
-      await client.query(
-        `
-        DELETE FROM notifications
-        WHERE team_id = $1
-          AND reference_type IN ('team_member', 'team_application', 'team_invitation')
-        `,
-        [team.teamId],
-      );
-      await client.query(
-        `
-        DELETE FROM team_vacant_role_tags
-        WHERE role_id IN (
-          SELECT id FROM team_vacant_roles WHERE team_id = $1
-        )
-        `,
-        [team.teamId],
-      );
-      await client.query(
-        `
-        DELETE FROM team_vacant_role_badges
-        WHERE role_id IN (
-          SELECT id FROM team_vacant_roles WHERE team_id = $1
-        )
-        `,
-        [team.teamId],
-      );
-      await client.query("DELETE FROM team_vacant_roles WHERE team_id = $1", [
-        team.teamId,
-      ]);
-      await client.query("DELETE FROM team_tags WHERE team_id = $1", [team.teamId]);
-      await client.query("DELETE FROM team_members WHERE team_id = $1", [
-        team.teamId,
-      ]);
-      await client.query("DELETE FROM teams WHERE id = $1", [team.teamId]);
-
-      deletedTeamIds.add(team.teamId);
-    }
-
-    const successorCandidatesByTeam = await getSuccessorCandidatesByTeam(
-      client,
-      teamsToTransfer.map((team) => team.teamId),
-      userId,
-    );
-
-    for (const team of teamsToTransfer) {
-      const overrideSuccessorId = ownershipOverrideMap.get(team.teamId);
-      const candidates = successorCandidatesByTeam.get(team.teamId) || [];
-
-      let successor = null;
-
-      if (overrideSuccessorId !== undefined) {
-        successor = candidates.find(
-          (candidate) => candidate.userId === overrideSuccessorId,
-        );
-
-        if (!successor) {
-          return rollbackAndRespond(400, {
-            success: false,
-            message: `Invalid ownership override for team ${team.teamName}`,
-          });
-        }
-      } else {
-        successor = candidates[0] || null;
-      }
-
-      if (!successor) {
-        throw new Error(`No successor candidate found for team ${team.teamId}`);
-      }
-
-      await client.query(
-        `
-        UPDATE team_members
-        SET role = 'owner'
-        WHERE team_id = $1
-          AND user_id = $2
-        `,
-        [team.teamId, successor.userId],
-      );
-
-      await client.query(
-        `
-        UPDATE teams
-        SET owner_id = $1
-        WHERE id = $2
-        `,
-        [successor.userId, team.teamId],
-      );
-
-      await insertNotificationRecord(client, {
-        userId: successor.userId,
-        type: "ownership_transferred",
-        title: `You are now the owner of ${team.teamName}`,
-        referenceType: "team_member",
-        referenceId: team.teamId,
-        teamId: team.teamId,
-        actorId: userId,
-      });
-
-      await client.query(
-        `INSERT INTO messages (sender_id, team_id, content, sent_at)
-         VALUES ($1, $2, $3, NOW())`,
-        [
-          null,
-          team.teamId,
-          `👑 OWNERSHIP_TEAM: ${DELETED_USER_DISPLAY_NAME} | ${successor.name}`,
-        ],
-      );
-
-      ownershipTransferEvents.push({
-        successorId: successor.userId,
-        teamId: team.teamId,
-      });
-    }
-
-    logDeletionPhase("Phase D - role and reference cleanup", {
-      filledRoleCount: filledRoles.length,
-    });
-
-    const reopenedRoles = filledRoles.filter(
-      (role) => !deletedTeamIds.has(role.teamId),
-    );
-
-    await client.query(
-      `
-      UPDATE team_vacant_roles
-      SET status = 'open',
-          filled_by = NULL,
-          updated_at = NOW()
-      WHERE filled_by = $1
-      `,
-      [userId],
-    );
-
-    if (reopenedRoles.length > 0) {
-      const reopenedTeamIds = Array.from(
-        new Set(reopenedRoles.map((role) => role.teamId)),
-      );
-
-      const roleRecipientsResult = await client.query(
-        `
-        SELECT team_id, user_id
-        FROM team_members
-        WHERE team_id = ANY($1::int[])
-          AND role IN ('owner', 'admin')
-          AND user_id != $2
-        ORDER BY team_id ASC, user_id ASC
-        `,
-        [reopenedTeamIds, userId],
-      );
-
-      const roleRecipientsByTeamId = new Map();
-
-      for (const row of roleRecipientsResult.rows) {
-        const teamId = Number(row.team_id);
-        const currentRecipients = roleRecipientsByTeamId.get(teamId) || [];
-        currentRecipients.push(Number(row.user_id));
-        roleRecipientsByTeamId.set(teamId, currentRecipients);
-      }
-
-      for (const role of reopenedRoles) {
-        await client.query(
-          `INSERT INTO messages (sender_id, team_id, content, sent_at)
-           VALUES ($1, $2, $3, NOW())`,
-          [null, role.teamId, `🔓 The role ${role.roleName} is now open again.`],
-        );
-
-        const recipients = roleRecipientsByTeamId.get(role.teamId) || [];
-
-        for (const recipientId of recipients) {
-          await insertNotificationRecord(client, {
-            userId: recipientId,
-            type: "role_reopened",
-            title: `The role ${role.roleName} is now open again in ${role.teamName}`,
-            teamId: role.teamId,
-            actorId: userId,
-          });
-        }
-
-        reopenedRoleEvents.push({
-          teamId: role.teamId,
-        });
-      }
-    }
-
-    await client.query(
-      `UPDATE team_vacant_roles SET created_by = NULL WHERE created_by = $1`,
-      [userId],
-    );
-    await client.query(
-      `UPDATE team_applications SET reviewed_by = NULL WHERE reviewed_by = $1`,
-      [userId],
-    );
-    await client.query(
-      `UPDATE user_badges SET awarded_by = NULL WHERE awarded_by = $1`,
-      [userId],
-    );
-    await client.query(`UPDATE tags SET created_by = NULL WHERE created_by = $1`, [
-      userId,
-    ]);
-    await client.query(
-      `UPDATE messages SET deleted_by = NULL WHERE deleted_by = $1`,
-      [userId],
-    );
-    await client.query(
-      `
-      UPDATE notifications
-      SET reference_id = NULL
-      WHERE actor_id = $1
-        AND reference_type IN ('team_invitation', 'team_application', 'badge_award')
-      `,
-      [userId],
-    );
-
-    logDeletionPhase("Phase E - delete user row", { userId });
-
-    await client.query("DELETE FROM users WHERE id = $1", [userId]);
-    await client.query("COMMIT");
-    transactionOpen = false;
-
-    logDeletionPhase("Phase F - post-transaction cleanup", {
-      teamEventCount: teamIdsForSockets.length,
-      dmPartnerCount: dmPartnerIds.length,
-    });
-
-    try {
-      if (avatarUrl || avatarFileId) {
-        await deleteImageKitFile(avatarUrl, avatarFileId);
-      }
-
-      const io =
-        req.app && typeof req.app.get === "function" ? req.app.get("io") : null;
-
-      if (io) {
-        for (const teamId of teamIdsForSockets) {
-          io.to(`team:${teamId}`).emit("team:member_left", { teamId, userId });
-        }
-
-        for (const partnerId of dmPartnerIds) {
-          io.to(`user:${partnerId}`).emit("conversation:deleted", {
-            partnerId: userId,
-          });
-        }
-
-        for (const event of ownershipTransferEvents) {
-          io.to(`user:${event.successorId}`).emit("notification:new", {
-            type: "ownership_transferred",
-            teamId: event.teamId,
-          });
-        }
-
-        for (const event of reopenedRoleEvents) {
-          io.to(`team:${event.teamId}`).emit("notification:new", {
-            type: "role_reopened",
-            teamId: event.teamId,
-          });
-        }
-      }
-    } catch (postCommitError) {
-      console.error("Error during post-transaction user deletion cleanup:", postCommitError);
-    }
-
-    return res.status(200).json({
-      success: true,
-      message: "Account deleted successfully",
-    });
-  } catch (error) {
-    if (transactionOpen) {
-      try {
-        await client.query("ROLLBACK");
-      } catch (rollbackError) {
-        console.error("Error rolling back user deletion:", rollbackError);
-      }
-    }
-
-    console.error("Error deleting user:", error);
-    res.status(500).json({
-      success: false,
-      message: "Error deleting user",
-      ...(process.env.NODE_ENV === "development" && { error: error.message }),
-    });
-  } finally {
-    if (client) {
-      client.release();
-    }
-  }
-};
-
-/**
- * @description Preview the impact of deleting a user's account
- * @route POST /api/users/:id/deletion-preview
- * @access Private (Requires authentication and password verification)
- */
-const deletionPreview = async (req, res) => {
-  try {
-    const userId = parseInt(req.params.id, 10);
-    const { password } = req.body;
-
-    if (Number(req.user.id) !== userId) {
-      return res.status(403).json({
-        success: false,
-        message: "You can only preview deletion for your own account",
-      });
-    }
-
-    if (!password) {
-      return res.status(400).json({
-        success: false,
-        message: "Password is required",
-      });
-    }
-
-    const userResult = await pool.query(
-      "SELECT id, password_hash FROM users WHERE id = $1",
-      [userId],
-    );
-
-    if (userResult.rows.length === 0) {
-      return res.status(404).json({
-        success: false,
-        message: "User not found",
-      });
-    }
-
-    const passwordMatches = await bcrypt.compare(
-      password,
-      userResult.rows[0].password_hash,
-    );
-
-    if (!passwordMatches) {
-      return res.status(401).json({
-        success: false,
-        message: "Password is incorrect",
-      });
-    }
-
-    const [
-      ownedTeamsResult,
-      rolesToReopenResult,
-      badgeAwardsGivenResult,
-      teamMembershipsResult,
-      directMessagesResult,
-    ] = await Promise.all([
-      pool.query(
-        `
-        SELECT
-          t.id AS team_id,
-          t.name AS team_name,
-          COUNT(tm_all.user_id)::int AS member_count,
-          (COUNT(tm_all.user_id) FILTER (WHERE tm_all.user_id != $1))::int AS other_member_count
-        FROM teams t
-        JOIN team_members tm_owner
-          ON tm_owner.team_id = t.id
-         AND tm_owner.user_id = $1
-         AND tm_owner.role = 'owner'
-        JOIN team_members tm_all ON tm_all.team_id = t.id
-        GROUP BY t.id, t.name
-        ORDER BY t.name ASC, t.id ASC
-        `,
-        [userId],
-      ),
-      pool.query(
-        `
-        SELECT
-          vr.id AS role_id,
-          vr.role_name,
-          vr.team_id,
-          t.name AS team_name
-        FROM team_vacant_roles vr
-        JOIN teams t ON t.id = vr.team_id
-        WHERE vr.filled_by = $1
-          AND vr.status = 'filled'
-        ORDER BY t.name ASC, vr.role_name ASC, vr.id ASC
-        `,
-        [userId],
-      ),
-      pool.query(
-        `
-        SELECT COUNT(*)::int AS count
-        FROM badge_awards
-        WHERE awarded_by_user_id = $1
-        `,
-        [userId],
-      ),
-      pool.query(
-        `
-        SELECT COUNT(*)::int AS count
-        FROM team_members
-        WHERE user_id = $1
-        `,
-        [userId],
-      ),
-      pool.query(
-        `
-        SELECT COUNT(*)::int AS count
-        FROM messages
-        WHERE (sender_id = $1 OR receiver_id = $1)
-          AND team_id IS NULL
-        `,
-        [userId],
-      ),
-    ]);
-
-    const ownedTeams = ownedTeamsResult.rows;
-    const teamIdsToTransfer = ownedTeams
-      .filter((team) => Number(team.other_member_count) > 0)
-      .map((team) => Number(team.team_id));
-
-    let successorByTeamId = new Map();
-
-    if (teamIdsToTransfer.length > 0) {
-      const successorsResult = await pool.query(
-        `
-        SELECT
-          ranked.team_id,
-          ranked.user_id,
-          ranked.role,
-          ranked.joined_at,
-          ranked.first_name,
-          ranked.last_name,
-          ranked.username
-        FROM (
-          SELECT
-            tm.team_id,
-            tm.user_id,
-            tm.role,
-            tm.joined_at,
-            u.first_name,
-            u.last_name,
-            u.username,
-            ROW_NUMBER() OVER (
-              PARTITION BY tm.team_id
-              ORDER BY
-                CASE
-                  WHEN tm.role = 'admin' THEN 0
-                  WHEN tm.role = 'member' THEN 1
-                  ELSE 2
-                END,
-                tm.joined_at ASC NULLS LAST,
-                tm.user_id ASC
-            ) AS row_number
-          FROM team_members tm
-          JOIN users u ON u.id = tm.user_id
-          WHERE tm.team_id = ANY($1::int[])
-            AND tm.user_id != $2
-            AND tm.role IN ('admin', 'member')
-        ) ranked
-        WHERE ranked.row_number = 1
-        `,
-        [teamIdsToTransfer, userId],
-      );
-
-      successorByTeamId = new Map(
-        successorsResult.rows.map((row) => [
-          Number(row.team_id),
-          {
-            userId: Number(row.user_id),
-            name: buildUserDisplayName(row),
-            role: row.role,
-            joinedAt: toIsoString(row.joined_at),
-          },
-        ]),
-      );
-    }
-
-    const teamsToDelete = [];
-    const teamsToTransfer = [];
-
-    for (const team of ownedTeams) {
-      const teamId = Number(team.team_id);
-      const memberCount = Number(team.member_count);
-      const otherMemberCount = Number(team.other_member_count);
-
-      if (otherMemberCount === 0) {
-        teamsToDelete.push({
-          teamId,
-          teamName: team.team_name,
-        });
-        continue;
-      }
-
-      const successor = successorByTeamId.get(teamId);
-
-      if (!successor) {
-        throw new Error(`No successor candidate found for team ${teamId}`);
-      }
-
-      teamsToTransfer.push({
-        teamId,
-        teamName: team.team_name,
-        successor,
-        memberCount,
-      });
-    }
-
-    const rolesToReopen = rolesToReopenResult.rows.map((row) => ({
-      roleId: Number(row.role_id),
-      roleName: row.role_name,
-      teamId: Number(row.team_id),
-      teamName: row.team_name,
-    }));
-
-    res.status(200).json({
-      success: true,
-      data: {
-        teamsToTransfer,
-        teamsToDelete,
-        rolesToReopen,
-        counts: {
-          badgeAwardsGiven: Number(badgeAwardsGivenResult.rows[0].count),
-          teamMemberships: Number(teamMembershipsResult.rows[0].count),
-          directMessages: Number(directMessagesResult.rows[0].count),
-        },
-      },
-    });
-  } catch (error) {
-    console.error("Error generating deletion preview:", error);
-    res.status(500).json({
-      success: false,
-      message: "Error generating deletion preview",
-      ...(process.env.NODE_ENV === "development" && { error: error.message }),
-    });
-  }
-};
-
-/**
- * @description Get tags for a specific user
- * @route GET /api/users/:id/tags
- * @access Public, with optional auth for own-profile hidden award visibility
- */
-const getUserTags = async (req, res) => {
-  try {
-    const userId = req.params.id;
-    const canViewHiddenAwards = Number(req.user?.id) === Number(userId);
-
-    const userVisibility = await pool.query(
-      `SELECT id, is_public, COALESCE(hide_badges, FALSE) AS hide_badges
-       FROM users
-       WHERE id = $1`,
-      [userId],
-    );
-
-    if (userVisibility.rows.length === 0) {
-      return res.status(404).json({ success: false, message: "User not found" });
-    }
-
-    const userRow = userVisibility.rows[0];
-    const userIsPublic =
-      userRow.is_public === true || userRow.is_public === "true";
-
-    if (!canViewHiddenAwards && !userIsPublic) {
-      let sharesTeam = false;
-      if (req.user) {
-        const teamCheck = await pool.query(
-          `SELECT 1 FROM team_members tm1
-           JOIN team_members tm2 ON tm1.team_id = tm2.team_id
-           WHERE tm1.user_id = $1 AND tm2.user_id = $2
-           LIMIT 1`,
-          [req.user.id, userId],
-        );
-        sharesTeam = teamCheck.rows.length > 0;
-      }
-      if (!sharesTeam) {
-        return res.status(404).json({ success: false, message: "User not found" });
-      }
-    }
-
-    await ensureBadgeVisibilityColumns();
-
-    const result = await pool.query(
-      `
-      SELECT
-        t.id,
-        t.name,
-        t.category,
-        t.supercategory,
-        ut.experience_level,
-        ut.interest_level,
-        COALESCE(tag_award_stats.badge_credits, 0)::INT AS badge_credits,
-        tag_award_stats.dominant_badge_category,
-        COALESCE(tag_award_stats.linked_badge_count, 0)::INT AS linked_badge_count,
-        COALESCE(tag_award_stats.awarder_count, 0)::INT AS awarder_count
-      FROM user_tags ut
-      JOIN users u ON u.id = ut.user_id
-      JOIN tags t ON ut.tag_id = t.id
-      LEFT JOIN LATERAL (
-        SELECT
-          COALESCE(SUM(ba.credits), 0)::INT AS badge_credits,
-          COUNT(*)::INT AS linked_badge_count,
-          COUNT(DISTINCT ba.awarded_by_user_id)::INT AS awarder_count,
-          (
-            SELECT b2.category
-            FROM badge_awards ba2
-            JOIN badges b2 ON b2.id = ba2.badge_id
-            WHERE ba2.tag_id = t.id
-              AND ba2.awarded_to_user_id = ut.user_id
-              AND (
-                $2::BOOLEAN = TRUE
-                OR COALESCE(u.hide_badges, FALSE) = TRUE
-                OR NOT (ba2.id = ANY(COALESCE(u.hidden_award_ids, '{}'::INTEGER[])))
-              )
-            GROUP BY b2.category
-            ORDER BY SUM(ba2.credits) DESC, b2.category ASC
-            LIMIT 1
-          ) AS dominant_badge_category
-        FROM badge_awards ba
-        WHERE ba.tag_id = t.id
-          AND ba.awarded_to_user_id = ut.user_id
-          AND (
-            $2::BOOLEAN = TRUE
-            OR COALESCE(u.hide_badges, FALSE) = TRUE
-            OR NOT (ba.id = ANY(COALESCE(u.hidden_award_ids, '{}'::INTEGER[])))
-          )
-      ) tag_award_stats ON TRUE
-      WHERE ut.user_id = $1
-    `,
-      [userId, canViewHiddenAwards],
-    );
-
-    res.status(200).json({
-      success: true,
-      data: userRow.hide_badges && !canViewHiddenAwards
-        ? result.rows.map((row) => ({
-            ...row,
-            badge_credits: 0,
-            dominant_badge_category: null,
-            linked_badge_count: 0,
-            awarder_count: 0,
-          }))
-        : result.rows,
-    });
-  } catch (error) {
-    console.error("Error fetching user tags:", error);
-    res.status(500).json({
-      success: false,
-      message: "Error fetching user tags",
-      ...(process.env.NODE_ENV === "development" && { error: error.message }),
-    });
-  }
-};
-
-/**
- * @description Update tags for a specific user
- * @route PUT /api/users/:id/tags
- * @access Private
- */
-const updateUserTags = async (req, res) => {
-  const client = await pool.connect();
-
-  try {
-    const userId = req.params.id;
-    const { tags } = req.body;
-
-    // Verify the user making the request is the same as the user being updated
-    if (req.user.id !== parseInt(userId)) {
-      return res.status(403).json({
-        success: false,
-        message: "You can only update your own tags",
-      });
-    }
-
-    await client.query("BEGIN");
-
-    // Calculate badge credits from badge_awards (source of truth)
-    // This works for both preserved AND re-added tags
-    const badgeCreditData = await client.query(
-      `SELECT
-         tag_id,
-         SUM(credits) AS badge_credits
-       FROM badge_awards
-       WHERE awarded_to_user_id = $1 AND tag_id IS NOT NULL
-       GROUP BY tag_id`,
-      [userId],
-    );
-    const creditMap = {};
-    for (const row of badgeCreditData.rows) {
-      creditMap[row.tag_id] = { badge_credits: Number(row.badge_credits) };
-    }
-
-    // Get dominant badge category per tag
-    const dominantData = await client.query(
-      `SELECT tag_id, badge_category
-       FROM v_user_tag_dominant_category
-       WHERE user_id = $1`,
-      [userId],
-    );
-    for (const row of dominantData.rows) {
-      if (creditMap[row.tag_id]) {
-        creditMap[row.tag_id].dominant_badge_category = row.badge_category;
-      }
-    }
-
-    // Delete existing tags for this user
-    await client.query("DELETE FROM user_tags WHERE user_id = $1", [userId]);
-
-    // Insert new tags
-    if (tags && tags.length > 0) {
-      const tagInserts = tags.map((tag) =>
-        client.query(
-          `
-          INSERT INTO user_tags (user_id, tag_id, experience_level, interest_level, badge_credits, dominant_badge_category)
-          VALUES ($1, $2, $3, $4, $5, $6)
-        `,
-          [
-            userId,
-            tag.tag_id || tag.id,
-            tag.experience_level || 2,
-            tag.interest_level || 3,
-            creditMap[tag.tag_id || tag.id]?.badge_credits || 0,
-            creditMap[tag.tag_id || tag.id]?.dominant_badge_category || null,
-          ],
-        ),
-      );
-
-      await Promise.all(tagInserts);
-    }
-
-    await client.query("COMMIT");
-
-    // Fetch the updated tags
-    const result = await pool.query(
-      `
-     SELECT 
-  t.id,
-  t.name,
-  t.category,
-  t.supercategory,
-  ut.experience_level,
-  ut.interest_level,
-  ut.badge_credits,
-  ut.dominant_badge_category,
-  (SELECT COUNT(*) FROM badge_awards ba WHERE ba.tag_id = t.id AND ba.awarded_to_user_id = ut.user_id) AS linked_badge_count,
-  (SELECT COUNT(DISTINCT ba.awarded_by_user_id) FROM badge_awards ba WHERE ba.tag_id = t.id AND ba.awarded_to_user_id = ut.user_id) AS awarder_count
-FROM user_tags ut
-JOIN tags t ON ut.tag_id = t.id
-WHERE ut.user_id = $1
-    `,
-      [userId],
-    );
-
-    res.status(200).json({
-      success: true,
-      message: "Tags updated successfully",
-      data: result.rows,
-    });
-  } catch (error) {
-    await client.query("ROLLBACK");
-    console.error("Error updating user tags:", error);
-    res.status(500).json({
-      success: false,
-      message: "Error updating user tags",
-      ...(process.env.NODE_ENV === "development" && { error: error.message }),
-    });
-  } finally {
-    client.release();
-  }
-};
-
-const updateUserBadgeVisibility = async (req, res) => {
-  try {
-    const userId = Number(req.params.id);
-    const awardId = Number(req.params.awardId);
-    const hidden = req.body?.hidden !== false;
-
-    if (!Number.isFinite(userId) || !Number.isFinite(awardId)) {
-      return res.status(400).json({
-        success: false,
-        message: "Valid user ID and award ID are required",
-      });
-    }
-
-    if (Number(req.user.id) !== userId) {
-      return res.status(403).json({
-        success: false,
-        message: "You can only update badge visibility on your own profile",
-      });
-    }
-
-    await ensureBadgeVisibilityColumns();
-
-    const awardResult = await pool.query(
-      `SELECT id, badge_id
-       FROM badge_awards
-       WHERE id = $1
-         AND awarded_to_user_id = $2
-       LIMIT 1`,
-      [awardId, userId],
-    );
-
-    if (awardResult.rows.length === 0) {
-      return res.status(404).json({
-        success: false,
-        message: "Badge award not found",
-      });
-    }
-
-    const result = await pool.query(
-      hidden
-        ? `UPDATE users
-           SET hidden_award_ids = (
-                 SELECT ARRAY(
-                   SELECT DISTINCT value
-                   FROM unnest(COALESCE(hidden_award_ids, '{}'::INTEGER[]) || $2::INTEGER) AS hidden_ids(value)
-                   ORDER BY value
-                 )
-               ),
-               updated_at = NOW()
-           WHERE id = $1
-           RETURNING hidden_award_ids`
-        : `UPDATE users
-           SET hidden_award_ids = array_remove(COALESCE(hidden_award_ids, '{}'::INTEGER[]), $2::INTEGER),
-               updated_at = NOW()
-           WHERE id = $1
-           RETURNING hidden_award_ids`,
-      [userId, awardId],
-    );
-
-    res.status(200).json({
-      success: true,
-      message: hidden
-        ? "Badge hidden successfully"
-        : "Badge made visible successfully",
-      data: {
-        awardId,
-        badgeId: awardResult.rows[0].badge_id,
-        hidden,
-        hiddenAwardIds: result.rows[0]?.hidden_award_ids ?? [],
-      },
-    });
-  } catch (error) {
-    console.error("Error updating badge visibility:", error);
-    res.status(500).json({
-      success: false,
-      message: "Error updating badge visibility",
-      ...(process.env.NODE_ENV === "development" && { error: error.message }),
-    });
-  }
-};
-
-/**
- * @description Get badges for a specific user
- * @route GET /api/users/:id/badges
- * @access Public
- */
-const getUserBadges = async (req, res) => {
-  try {
-    const userId = req.params.id;
-    const canViewHiddenAwards = Number(req.user?.id) === Number(userId);
-
-    const userVisibility = await pool.query(
-      `SELECT id, is_public, COALESCE(hide_badges, FALSE) AS hide_badges
-       FROM users
-       WHERE id = $1`,
-      [userId],
-    );
-
-    if (userVisibility.rows.length === 0) {
-      return res.status(404).json({ success: false, message: "User not found" });
-    }
-
-    const userRow = userVisibility.rows[0];
-    const userIsPublic =
-      userRow.is_public === true || userRow.is_public === "true";
-
-    if (!canViewHiddenAwards && !userIsPublic) {
-      let sharesTeam = false;
-      if (req.user) {
-        const teamCheck = await pool.query(
-          `SELECT 1 FROM team_members tm1
-           JOIN team_members tm2 ON tm1.team_id = tm2.team_id
-           WHERE tm1.user_id = $1 AND tm2.user_id = $2
-           LIMIT 1`,
-          [req.user.id, userId],
-        );
-        sharesTeam = teamCheck.rows.length > 0;
-      }
-      if (!sharesTeam) {
-        return res.status(404).json({ success: false, message: "User not found" });
-      }
-    }
-
-    if (!canViewHiddenAwards && userRow.hide_badges) {
-      return res.status(200).json({ success: true, data: [] });
-    }
-
-    await ensureBadgeVisibilityColumns();
-
-    const result = await pool.query(
-      `
-      SELECT
-        ba.id AS award_id,
-
-        -- badge fields
-        b.id AS badge_id,
-        b.name AS badge_name,
-        b.description AS badge_description,
-        b.category AS badge_category,
-        b.image_url AS badge_image_url,
-        b.color AS badge_color,
-        b.cat_image_url AS badge_category_image_url,
-
-        -- award fields
-        ba.credits,
-        ba.created_at AS awarded_at,
-        ba.reason,
-        ba.context_type,
-        ba.context_id,
-        ba.team_id,
-        ba.custom_team_name,
-        ba.project_name,
-        tag.name AS tag_name,
-        tag.category AS tag_category,
-        COALESCE(t.name, ba.custom_team_name) AS team_name,
-
-        -- awarder fields
-        ba.awarded_by_user_id AS awarded_by_user_id,
-        awarder.username AS awarded_by_username,
-        awarder.first_name AS awarded_by_first_name,
-        awarder.last_name AS awarded_by_last_name,
-        awarder.avatar_url AS awarded_by_avatar_url,
-        awarder.is_synthetic AS awarded_by_is_synthetic
-
-      FROM badge_awards ba
-      JOIN badges b ON ba.badge_id = b.id
-      LEFT JOIN users awarder ON ba.awarded_by_user_id = awarder.id
-      LEFT JOIN teams t ON ba.team_id = t.id
-      LEFT JOIN tags tag ON ba.tag_id = tag.id
-      LEFT JOIN users awardee ON awardee.id = ba.awarded_to_user_id
-      WHERE ba.awarded_to_user_id = $1
-        AND (
-          $2::BOOLEAN = TRUE
-          OR NOT (ba.id = ANY(COALESCE(awardee.hidden_award_ids, '{}'::INTEGER[])))
-        )
-      ORDER BY ba.created_at DESC, ba.id DESC
-      `,
-      [userId, canViewHiddenAwards],
-    );
-
-    res.status(200).json({ success: true, data: result.rows });
-  } catch (error) {
-    console.error("Error fetching user badges:", error);
-    res.status(500).json({
-      success: false,
-      message: "Error fetching user badges",
-      ...(process.env.NODE_ENV === "development" && { error: error.message }),
-    });
-  }
-};
-
-// ============================================================
-// USER BLOCKS
-// ============================================================
-
-// Guards that the authenticated user is acting on their own blocklist.
-const isSelf = (req) => Number(req.user?.id) === Number(req.params.id);
-
-// Tell both parties' clients (all devices) to re-sync block state in realtime,
-// so a block/unblock instantly hides/restores chats without a manual refresh.
-const emitBlocksUpdated = (req, blockerId, blockedId, blocked) => {
-  const io = req.app.get("io");
-  if (!io) return;
-  io.to(`user:${blockedId}`).emit("blocks:updated", {
-    withUserId: Number(blockerId),
-    blocked,
-  });
-  io.to(`user:${blockerId}`).emit("blocks:updated", {
-    withUserId: Number(blockedId),
-    blocked,
-  });
-};
-
-/**
- * @description List users the current user has blocked.
- * @route GET /api/users/:id/blocks
- * @access Private (self only)
- */
-const getBlockedUsers = async (req, res) => {
-  try {
-    if (!isSelf(req)) {
-      return res.status(403).json({ success: false, message: "Forbidden" });
-    }
-    const blocked = await userModel.getBlockedUsers(req.user.id);
-    res.status(200).json({ success: true, data: blocked });
-  } catch (error) {
-    console.error("Error fetching blocked users:", error);
-    res.status(500).json({
-      success: false,
-      message: "Error fetching blocked users",
-      ...(process.env.NODE_ENV === "development" && { error: error.message }),
-    });
-  }
-};
-
-/**
- * @description Block a user.
- * @route POST /api/users/:id/blocks
- * @access Private (self only)
- */
-const blockUser = async (req, res) => {
-  try {
-    if (!isSelf(req)) {
-      return res.status(403).json({ success: false, message: "Forbidden" });
-    }
-    const blockedId = Number(req.body.blocked_id ?? req.body.blockedId);
-    if (!blockedId || Number.isNaN(blockedId)) {
-      return res
-        .status(400)
-        .json({ success: false, message: "blockedId is required" });
-    }
-    if (blockedId === Number(req.user.id)) {
-      return res
-        .status(400)
-        .json({ success: false, message: "You cannot block yourself" });
-    }
-
-    const target = await pool.query(`SELECT id FROM users WHERE id = $1`, [
-      blockedId,
-    ]);
-    if (target.rows.length === 0) {
-      return res
-        .status(404)
-        .json({ success: false, message: "User not found" });
-    }
-
-    await userModel.blockUser(req.user.id, blockedId);
-    emitBlocksUpdated(req, req.user.id, blockedId, true);
-    res
-      .status(201)
-      .json({ success: true, message: "User blocked successfully" });
-  } catch (error) {
-    console.error("Error blocking user:", error);
-    res.status(500).json({
-      success: false,
-      message: "Error blocking user",
-      ...(process.env.NODE_ENV === "development" && { error: error.message }),
-    });
-  }
-};
-
-/**
- * @description Unblock a user.
- * @route DELETE /api/users/:id/blocks/:blockedId
- * @access Private (self only)
- */
-const unblockUser = async (req, res) => {
-  try {
-    if (!isSelf(req)) {
-      return res.status(403).json({ success: false, message: "Forbidden" });
-    }
-    const blockedId = Number(req.params.blockedId);
-    if (!blockedId || Number.isNaN(blockedId)) {
-      return res
-        .status(400)
-        .json({ success: false, message: "Invalid blocked user id" });
-    }
-    await userModel.unblockUser(req.user.id, blockedId);
-    emitBlocksUpdated(req, req.user.id, blockedId, false);
-    res
-      .status(200)
-      .json({ success: true, message: "User unblocked successfully" });
-  } catch (error) {
-    console.error("Error unblocking user:", error);
-    res.status(500).json({
-      success: false,
-      message: "Error unblocking user",
-      ...(process.env.NODE_ENV === "development" && { error: error.message }),
-    });
-  }
-};
-
-/**
- * @description Get every user id in a block relationship with the current user
- *              (either direction) — used by the client to mutually anonymize
- *              blocked users in shared teams.
- * @route GET /api/users/:id/block-relationships
- * @access Private (self only)
- */
-const getBlockRelationships = async (req, res) => {
-  try {
-    if (!isSelf(req)) {
-      return res.status(403).json({ success: false, message: "Forbidden" });
-    }
-    const ids = await userModel.getBlockRelationshipIds(req.user.id);
-    res.status(200).json({ success: true, data: { ids } });
-  } catch (error) {
-    console.error("Error fetching block relationships:", error);
-    res.status(500).json({
-      success: false,
-      message: "Error fetching block relationships",
-      ...(process.env.NODE_ENV === "development" && { error: error.message }),
-    });
-  }
-};
-
 module.exports = {
   getUsers,
   getUserById,
   updateUser,
-  deletionPreview,
-  deleteUser,
   deleteAvatar,
-  getUserTags,
-  updateUserTags,
-  updateUserBadgeVisibility,
-  getUserBadges,
-  getBlockedUsers,
-  blockUser,
-  unblockUser,
-  getBlockRelationships,
 };
